Feature/chat

+++ app/src/main/java/com/aoztg/greengrim/data/model/request/CreateCertificationRequest.kt
@@ -1,7 +1,7 @@
 package com.aoztg.greengrim.data.model.request
 
 data class CreateCertificationRequest(
-    val challengeId: Int,
+    val challengeId: Long,
     val imgUrl: String,
     val description: String,
     val round: Int
