+++ app/src/main/java/com/aoztg/greengrim/data/model/request/CreateNftRequest.kt
@@ -2,7 +2,7 @@ package com.aoztg.greengrim.data.model.request
 
 data class CreateNftRequest(
     val password: String,
-    val grimId: Int,
+    val grimId: Long,
     val title: String,
     val description: String,
 )
