+++ app/src/main/java/com/aoztg/greengrim/data/model/request/SellNftRequest.kt
@@ -2,6 +2,6 @@ package com.aoztg.greengrim.data.model.request
 
 data class SellNftRequest(
     val payPassword: String,
-    val nftId: Int,
+    val nftId: Long,
     val price: Double
 )
