+++ app/src/main/java/com/aoztg/greengrim/data/model/response/ChatRoomsResponse.kt
@@ -1,8 +1,8 @@
 package com.aoztg.greengrim.data.model.response
 
 data class ChatRoomsResponse(
-    val chatroomId: Int,
-    val challengeId: Int,
+    val chatroomId: Long,
+    val challengeId: Long,
     val title: String,
     val afterDay: String,
     val imgUrl: String
