+++ app/src/main/java/com/aoztg/greengrim/data/model/response/CreateChallengeResponse.kt
@@ -1,7 +1,7 @@
 package com.aoztg.greengrim.data.model.response
 
 data class CreateChallengeResponse(
-    val challengeId: Int,
-    val chatroomId: Int,
+    val challengeId: Long,
+    val chatroomId: Long,
     val title: String
 )
