+++ app/src/main/java/com/aoztg/greengrim/presentation/chatmanager/model/UiUnReadChatData.kt
@@ -1,7 +1,7 @@
 package com.aoztg.greengrim.presentation.chatmanager.model
 
 data class UiUnReadChatData(
-    val chatId: Int = -1,
+    val chatId: Long = -1,
     val unReadCount: Int = 0,
     val recentChat: String = "",
     val recentChatTime: String = "",
