+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/home/model/UiHomeInfo.kt
@@ -1,7 +1,7 @@
 package com.aoztg.greengrim.presentation.ui.home.model
 
 data class UiHomeInfo(
-    val id: Int=0,
-    val nickName: String="",
-    val carbonReduction: String=""
+    val id: Long = 0,
+    val nickName: String = "",
+    val carbonReduction: String = ""
 )
