+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/market/model/UiGrimForNft.kt
@@ -1,8 +1,8 @@
 package com.aoztg.greengrim.presentation.ui.market.model
 
 data class UiGrimForNft(
-    val id: Int = -1,
+    val id: Long = -1,
     val image: String = "",
     val title: String = "",
-    val selectListener: (Int, String) -> Unit
+    val selectListener: (Long, String) -> Unit
 )
